✨ Convert .all NSEventMonitors to use PassiveEventMonitor
